完成faster rcnn 端到端训练过程

+++ faster_rcnn/utils/tf_utils.py
@@ -225,10 +225,11 @@ def pad_to_fixed_size_with_negative(input_tensor, fixed_size, negative_num):
     input_size = tf.shape(input_tensor)[0]
     # tag 列 padding
     positive_num = input_size - negative_num
-    column_padding = tf.concat([tf.ones([positive_num], dtype=tf.int64),
-                                tf.ones([negative_num], dtype=tf.int64) * -1],
+    column_padding = tf.concat([tf.ones([positive_num]),
+                                tf.ones([negative_num]) * -1],
                                axis=0)
-    x = tf.concat([input_tensor, tf.expand_dims(column_padding, axis=1)], axis=1)
+    # 都转为float,拼接
+    x = tf.concat([tf.cast(input_tensor, tf.float32), tf.expand_dims(column_padding, axis=1)], axis=1)
     # 行padding
     padding_size = tf.maximum(0, fixed_size - input_size)
     x = tf.pad(x, [[0, padding_size], [0, 0]], mode='CONSTANT', constant_values=0)
